+++ events/dialogController.js
@@ -1,14 +1,22 @@
 const dialog = document.querySelector("dialog");
 
 const openDialog = async (e) => {
-  const { url } = e.currentTarget.dataset;
+  const { url, formAction } = e.currentTarget.dataset;
   console.log("e.currentTarget:", e.currentTarget);
   console.log("e.currentTarget.dataset:", e.currentTarget.dataset);
   console.log("url:", url);
-  const form = await fetchForm(url);
-  console.log(form);
+  const content = await fetchContent(url, formAction);
+  console.log(content);
   console.log("window.location:", window.location);
-  dialog.append(form);
+  // Need to set the form's action attribute
+  // /drive/folder/:folderID/files/upload
+  // /drive/files/upload
+  // Issue
+  //  Invalid form will render a new form on POST
+  // content.setAttribute("action", formAction);
+  // console.log("form.action:", form.action);
+  // console.log("form.getAttribute('action'):", form.getAttribute("action"));
+  dialog.append(content);
   dialog.showModal();
 };
 
@@ -19,8 +27,7 @@ const closeDialog = (e) => {
 
 const onCloseHandler = () => {
   console.log("onCloseHandler running...");
-  const form = dialog.querySelector("form");
-  form.remove();
+  dialog.firstChild.remove();
 };
 
 // stopPropagation on the dialog element's child/children element(s)
@@ -29,7 +36,7 @@ const stopPropagation = (e) => {
   e.stopPropagation();
 };
 
-const fetchForm = async (url) => {
+const fetchContent = async (url, formAction) => {
   const fetchURL = `/components/${url}`;
   return await fetch(fetchURL, {
     method: "GET",
@@ -40,7 +47,16 @@ const fetchForm = async (url) => {
     .then((html) => {
       const parser = new DOMParser();
       const doc = parser.parseFromString(html, "text/html");
-      return doc.querySelector("form");
+      console.log("--------------------------------------------");
+      console.log(doc);
+      if (formAction) {
+        const form = doc.querySelector("form");
+        form.setAttribute("action", formAction);
+        return form;
+      }
+      console.log("doc.body:", doc.body.firstElementChild);
+
+      return doc.querySelector("article");
     })
     .catch((err) => {
       // Is there a proper way to handle errors when fetching components?